requirements.txt
dill==0.3.3
tensorflow==2.5.0
tensorflow-hub==0.12.0
wordcloud==1.8.1
Pillow==8.1.2
plotly==4.14.3
requests==2.25.1
retrying==1.3.3
flask==1.1.0
gunicorn==20.0.4
requests==2.25.1
pandas==1.2.3
numpy==1.21.2